TowerBuyButtonUI script added
Added Logic for the TowerBuyButton so that it dynamically gets and sets the information for the given Tower.
GameManager moved in the Load order so that the GameManager is the very first script initialized in the Scene.
See Project Settings>Script Execution Order for details.
TowerButtonUI subscribes to onPlayerCashChanged Event from teh GameManager.