Bugfix
Game can now NOT start the next round if there are still enemies to spawn, even if there are no enemies on the field.

This is done via the WaveSpawner.waveEnemies integer that is set when a new wave is being spawned. This integer add together alle the spawnamounts for all the enemysets in the wave. When an Enemy i destroyed there happens a subtraction from waveenemies to reflect that there is one less enemy to destroy befor the round/wave is won.